requirements.txt
keyboard~=0.13.5
colorama~=0.4.6
pyperclip~=1.8.2
PyAutoGUI~=0.9.53
mouse~=0.7.1
pyttsx3~=2.90
Pillow~=9.3.0
termcolor~=2.1.0
requests~=2.28.1
moviepy~=1.0.3
rotate-screen==0.1.0